+++ server/utils/utils.go
@@ -74,6 +74,7 @@ func extractParamsFromPost(c *gin.Context) (map[string]interface{}, error) {
 		"arch":                upReq.Arch,
 		"changelog":           upReq.Changelog,
 		"artifacts_to_delete": upReq.ArtifactsToDelete,
+		"updater":             upReq.Updater,
 	}, nil
 }